Renamed event metadata definition test for user login event
