+++ src/Core/RulesEvent.php
@@ -10,7 +10,7 @@ namespace Drupal\rules\Core;
 use Drupal\Core\Plugin\ContextAwarePluginBase;
 
 /**
- * Base class for rules actions.
+ * Base class for rules events.
  */
 class RulesEvent extends ContextAwarePluginBase implements RulesEventInterface {
